Pagination logic complete
Can now infinitely paginate from pages 3 - *. Pages 1 and 2 are hardcoded due to the nature
of the algo not applying to them. This is the best I could come up with! I have left a pagination logic example
in a comment just before the logic itself to clearly show what is happening.